// src/interfaces/persona-response.interface.ts
// Generated by https://quicktype.io

export interface ProfesorGrupo extends PersonaResponse{
    materia:string;
    grupo:string
}


export interface PersonaResponse {
    correo_institucional:  string;
    nombres:               string;
    apellidos:             string;
    cod_tipo_doc:          number | null;
    cod_rol:               number | null;
    codigo:                string;
    correo_personal?:      null | string;
    documento?:            null | string;
    celular?:              null | string;
    semestres?:            number | null;
    img?:                  string | null;
    perfil_completado:     boolean;
}